Fix moar tests
